examples/gl-connectors/google_docs/pyproject.toml
[project]
name = "google-docs"
version = "0.1.0"
description = "Google Docs connector examples"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "glaip-sdk[local]>=0.8.1",
    "gllm-tools-binary>=0.1.5,<0.2.0",
    "langchain",
    "langchain-openai",
]
